Added search function
Added search form in HTML. Added 2 new functions in JS to be able to search for city based on input provided by user.
While testing, realized the weather-icon or image that was being created did not overwrite the previous search results but added on top of it, changed the code to make it render properly.