Unit change enables
You can now interact with F and C to show the temperature of the location based on Fahrenheit and Celsius.
Adds or removes an "active" class that through CSS changes the format of the links. One bug I need to fix: if you search something and you are in Fahrenheit, the active class stays in Fahrenheit but it shows the weather in celsius. Will play with it to work on that.