Made Celsius Active default
Made celsius link active by default by adding the class "active" to celsius when someone searches and removing "active" from fahrenheit link. Ideally, if user clicks on Fahrenheit, I want the next search to be displayed in fahrenheit. Will work on that later.
Also need to work on displaying the time of the location rather than when it was last updated, i like it better like that.