Country names display fully. US city displays state not country
Added Javascript to show the country name (not iso Code) if they are outside of the US. If the city is within the US, it only shows the state and not the country. To do this, I did add another API from the same provides, GEOLOCATION API so be able to obtain the state information.