CSS improvement and temporary removal of Unit conversion
For contrast improvement, I decreased the transparency of the white filter on top of background image and made the forecast max temp thicker.
I also removed the unit conversion for the time being. I will be adding this feature again and applying it to both today's weather and the forecast. This will happen after I provide my code to SheCodes to complete Plus cert.